get_boot_time_us: check if there is a preempted systick interrupt pending
Reported by Simon Laube: The usage of get_boot_time_us() in the camera
interrupt handler caused issues where sometimes it would return a
timestamp that is 1ms in the past. The result is that some frames deltas
would be reported as 1ms too short while the next frame after such an
incident would be 1ms too long.