Change result accumulator to handle dropped frame time correctly
This only affects the legacy data in pixels / accumulation period.

Based on 04f365d by Simon Laube

The full_time variable of the accumulator did not include the time of
the dropped frames. This is fixed now. Accumulator calculations relying
on the full_time variable are now correct event when frames are dropped.
(for example when taking a snapshot)

The full_time variable is currently only used by the old flow message
which uses pixels as the output format. It is used to scale the measured
pixel values to appear like we did not drop any frames. This scaling is
correct if we assume that the speed remains constant during the
accumulation period.